requirements.txt
streamlit
openai>=1.0.0
anthropic
google-generativeai
supabase
extra-streamlit-components
streamlit-cookies-manager